backend/requirements.txt

Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-decouple==3.8